Error handled

+++ Operators.java
@@ -4,11 +4,11 @@ public class Operators {
         double x = 5;
         double y = 10;
 
-        int z1 = x + y;
-        int z2 = x - y;
+        double z1 = x + y;
+        double z2 = x - y;
         double z3 = x * y;
         double z4 = x / y;
-        int z5 = x % y;
+        double z5 = x % y;
 
         
         System.out.println(z1);
